# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CreatorHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List all links for current user
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a new link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
  /links/{id}:
    patch:
      operationId: updateLink
      tags: [links]
      summary: Update a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdate"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /links/reorder:
    patch:
      operationId: reorderLinks
      tags: [links]
      summary: Reorder links by providing ordered IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkReorder"
      responses:
        "200":
          description: Reordered links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"
  /links/{id}/click:
    post:
      operationId: trackLinkClick
      tags: [links]
      summary: Track a link click (public endpoint)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Click tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── PRODUCTS ────────────────────────────────────────────────────────────────
...
  /ai-tools/hooks:
    post:
      operationId: generateHooks
      tags: [ai_tools]
      summary: Generate viral hooks for TikTok/Reels/Shorts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HookGeneratorInput"
      responses:
        "200":
          description: Generated hooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HookGeneratorOutput"
  /ai-tools/script:
    post:
      operationId: generateScript
      tags: [ai_tools]
      summary: Generate a short-form video script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptGeneratorInput"
      responses:
        "200":
          description: Generated script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptGeneratorOutput"
  /ai-tools/thumbnail-analyze:
    post:
      operationId: analyzeThumbnail
      tags: [ai_tools]
      summary: Analyze a video title or thumbnail for virality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThumbnailAnalyzerInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThumbnailAnalyzerOutput"
  /ai-tools/trends:
    get:
      operationId: getTrends
      tags: [ai_tools]
      summary: Get trending topics by niche
      parameters:
        - name: niche
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Trending topics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendsOutput"

  # ─── UPLOADS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ─── PROFILE ────────────────────────────────────────────────────────────
...
    Link:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        url:
          type: string
        icon:
          type: ["string", "null"]
        clicks:
          type: integer
        sortOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - url
        - clicks
        - sortOrder
        - isActive
        - createdAt
    LinkInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        url:
          type: string
          minLength: 1
        icon:
          type: string
      required:
        - title
        - url
    LinkUpdate:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
    LinkReorder:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required:
        - ids
...
    HookGeneratorInput:
      type: object
      properties:
        topic:
          type: string
        niche:
          type: string
        platform:
          type: string
          enum: [tiktok, instagram, youtube, shorts]
        count:
          type: integer
      required:
        - topic
        - niche
    HookGeneratorOutput:
      type: object
      properties:
        hooks:
          type: array
          items:
            type: string
        platform:
          type: string
      required:
        - hooks
        - platform
    ScriptGeneratorInput:
      type: object
      properties:
        topic:
          type: string
        niche:
          type: string
        duration:
          type: integer
        platform:
          type: string
        tone:
          type: string
      required:
        - topic
        - niche
    ScriptGeneratorOutput:
      type: object
      properties:
        hook:
          type: string
        body:
          type: string
        cta:
          type: string
        fullScript:
          type: string
        estimatedDuration:
          type: integer
      required:
        - hook
        - body
        - cta
        - fullScript
        - estimatedDuration
    ThumbnailAnalyzerInput:
      type: object
      properties:
        title:
          type: string
        thumbnailUrl:
          type: string
        niche:
          type: string
      required:
        - title
    ThumbnailAnalyzerOutput:
      type: object
      properties:
        score:
          type: integer
        grade:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        optimizedTitle:
          type: string
      required:
        - score
        - grade
        - strengths
        - improvements
        - optimizedTitle
    TrendItem:
      type: object
      properties:
        topic:
          type: string
        niche:
          type: string
        engagement:
          type: string
        contentIdeas:
          type: array
          items:
            type: string
      required:
        - topic
        - niche
        - engagement
        - contentIdeas
    TrendsOutput:
      type: object
      properties:
        trends:
          type: array
          items:
            $ref: "#/components/schemas/TrendItem"
        generatedAt:
          type: string
          format: date-time
      required:
        - trends
        - generatedAt

    # ─── UPLOADS ──────────────────────────────────────────────────────────────
...